requirements.txt
autopep8==2.0.4
docker==6.1.3
Faker==19.9.0
Flask==2.3.3
hubspot-api-client==8.1.0
mysql-connector-python==8.1.0
pg==0.1
psycopg2==2.9.7
python-dateutil==2.8.2
python-dotenv==0.21.0
stripe==6.7.0
urllib3==2.0.4
websocket-client==1.6.3

